improve performance on intel graphics
