+++ game/graphics/opengl_renderer/SkyRenderer.h
@@ -3,35 +3,8 @@
 #include "game/graphics/opengl_renderer/BucketRenderer.h"
 #include "game/graphics/opengl_renderer/DirectRenderer.h"
 #include "game/graphics/opengl_renderer/tfrag/TFragment.h"
+#include "game/graphics//opengl_renderer/SkyBlendGPU.h"
 
-class SkyBlender {
- public:
-  SkyBlender();
-  ~SkyBlender();
-
-  struct Stats {
-    int sky_draws = 0;
-    int cloud_draws = 0;
-    int sky_blends = 0;
-    int cloud_blends = 0;
-  };
-
-  Stats do_sky_blends(DmaFollower& dma, SharedRenderState* render_state, ScopedProfilerNode& prof);
-
- private:
-  GLuint m_framebuffers[2];  // sky, clouds
-  GLuint m_textures[2];      // sky, clouds
-  int m_sizes[2] = {32, 64};
-  GLuint m_gl_vertex_buffer;
-
-  struct Vertex {
-    float x = 0;
-    float y = 0;
-    float intensity = 0;
-  };
-
-  Vertex m_vertex_data[6];
-};
 
 /*!
  * Handles texture blending for the sky.
@@ -41,7 +14,7 @@ class SkyBlendHandler : public BucketRenderer {
  public:
   SkyBlendHandler(const std::string& name,
                   BucketId my_id,
-                  std::shared_ptr<SkyBlender> shared_blender);
+                  std::shared_ptr<SkyBlendGPU> shared_gpu_blender);
   void render(DmaFollower& dma, SharedRenderState* render_state, ScopedProfilerNode& prof) override;
   void draw_debug_window() override;
 
@@ -50,8 +23,8 @@ class SkyBlendHandler : public BucketRenderer {
                          SharedRenderState* render_state,
                          ScopedProfilerNode& prof);
 
-  std::shared_ptr<SkyBlender> m_shared_blender;
-  SkyBlender::Stats m_stats;
+  std::shared_ptr<SkyBlendGPU> m_shared_gpu_blender;
+  SkyBlendGPU::Stats m_gpu_stats;
   TFragment m_tfrag_renderer;
 };
