+++ game/graphics/opengl_renderer/shaders/sky.frag
@@ -3,7 +3,7 @@
 out vec4 color;
 
 in vec4 fragment_color;
-in vec3 tex_coord;
+noperspective in vec3 tex_coord;
 uniform sampler2D tex_T0;
 
 void main() {
