+++ game/graphics/opengl_renderer/tfrag/Tfrag3.h
@@ -48,6 +48,9 @@ class Tfrag3 {
   struct TreeCache {
     tfrag3::TFragmentTreeKind kind;
     GLuint vertex_buffer = -1;
+    GLuint index_buffer = -1;
+    std::vector<u32> index_list;
+    GLuint time_of_day_texture;
     GLuint vao;
     u32 vert_count = 0;
     const std::vector<tfrag3::StripDraw>* draws = nullptr;
@@ -71,20 +74,14 @@ class Tfrag3 {
   struct Cache {
     std::vector<u8> vis_temp;
     std::vector<std::pair<int, int>> draw_idx_temp;
-    std::vector<u32> index_list;
   } m_cache;
 
   std::string m_level_name;
 
   std::vector<GLuint> m_textures;
   std::vector<TreeCache> m_cached_trees;
-  GLuint m_time_of_day_texture = -1;
-  bool m_has_time_of_day_texture = false;
 
   std::vector<math::Vector<u8, 4>> m_color_result;
-
-  bool m_has_index_buffer = false;
-  GLuint m_index_buffer = -1;
 
   GLuint m_debug_vao = -1;
   GLuint m_debug_verts = -1;
